requirements.txt
fastapi[standard]
python-dotenv
paypal-server-sdk
